разделил моки, исправил названия переменных, добавили логику отрисовки фильтров
