+++ src/render.js
@@ -27,5 +27,5 @@ export const renderElement = (container, element, place) => {
 export const createElement = (template) => {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
-  return newElement.firstChild;
+  return newElement.firstElementChild;
 };
