+++ src/test.js
@@ -1,0 +1,12 @@
+// const arra = [1,2,[3,4, [5,6]],7];
+// function flatten (array) {
+//   return array.reduce((acc, it) => {
+//     if(Array.isArray(it)){
+//       // return [...acc, ...flatten(it)];
+//       return acc.concat(flatten(it));
+//     }
+//     acc.push(it);
+//     return acc;
+//   }, []);
+// }
+// console.log(flatten(arra));
