+++ src/view/user-title-view.js
@@ -2,7 +2,7 @@ import { createElement } from '../render.js';
 
 const createUserTitleTemplate = (values) => (
   `<section class="header__profile profile">
-      <p class="profile__rating">${values.userTitle}</p>
+      <p class="profile__rating">${values.rating}</p>
       <img class="profile__avatar" src="${values.bitmap}" alt="Avatar" width="35" height="35">
     </section>`);
